Separated bishop movement logic to a different file
